uninst_requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
comm==0.1.4
cryptography==41.0.3
debugpy==1.6.7.post1
decorator==5.1.1
distro==1.8.0
duckdb==0.8.1
et-xmlfile==1.1.0
executing==1.2.0
idna==3.4
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
jupyter_client==8.3.0
jupyter_core==5.3.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
numpy==1.25.2
opencv-python==4.8.0.76
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
Pygments==2.16.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
stack-data==0.6.2
tornado==6.3.3
traitlets==5.9.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
